+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+* Fix excessive CPU usage caused by `threading.Event.wait`
+
+
 # 0.4.5
 
 * Expose `RunOnNotification` and `CallMiddleware`